integration test with comparison of output archive to ref

+++ python/snewpy/test/snowglobes_integration_benchmark_test.py
@@ -11,42 +11,27 @@ outdir = Path(snowglobes_dir)/'out'
 
 #simulation setup
 model_file = basedir/'models/Kuroda_2020/LnuR10B12.dat'
-model_type =model_file.parent.stem
+model_type = model_file.parent.stem
 xform_type = 'AdiabaticMSW_NMO'
+detectors= 'all'
 
 fluence_name = f'test_{model_type}_{xform_type}'
 fluence_file = model_file.parent/f'{fluence_name}.tar.bz2'
-detectors= 'all'
+output_archive = model_file.parent/f'{fluence_name}_SNOprocessed.tar.gz'
 
 #checks setup
-ref_dir = basedir/'out_ref1' 
+reference_archive = basedir/f'reference_{fluence_name}.tar.gz' 
 
 from snewpy import snowglobes
 
-#main steps defined here
-def generate():
-    f = snowglobes.generate_fluence(model_file, model_type, xform_type, d=10, 
-                                    output_filename=fluence_name)
-    return f
-
-def simulate():
-    #cleanup output directory first
-    for f in outdir.glob('*.*'):
-        f.unlink()
-    snowglobes.simulate(snowglobes_dir,fluence_file, detector_input=detectors)
-
-def collate():
-    #recreate the file to avoid crashes
-    Path(f'{snowglobes_dir}/fluxes/{fluence_name}.dat').touch()
-    try:
-        result = snowglobes.collate(snowglobes_dir,fluence_file, detector_input=detectors, 
-                                            skip_plots=True, remove_generated_files=False)
-    finally:
-        #return back to original directory
-        os.chdir(basedir)
-    return result
-
-def read_data(fname):
+#-------------------------------------------------
+def _cleanup(path, pattern='*.*'):
+    "Remove all the files from path/ matching pattern"
+    if path.exists():
+        for f in path.glob(pattern):
+            f.unlink()
+    
+def _read_data(fname):
     return np.loadtxt(fname,comments=['---','Total','#','Energy'], dtype='f8')
 
 def _check_file_in_directory(ref_file, check_dir):
@@ -61,6 +46,28 @@ def _check_file_in_directory(ref_file, check_dir):
         d1 = read_data(check_file)
         print(d0-d1)
         assert np.allclose(d0,d1)
+#--------------------------------------------
+#main steps defined here
+def generate():
+    f = snowglobes.generate_fluence(model_file, model_type, xform_type, d=10, 
+                                    output_filename=fluence_name)
+    return f
+
+def simulate():
+    #cleanup output directory first
+    _cleanup(outdir)
+    snowglobes.simulate(snowglobes_dir,fluence_file, detector_input=detectors)
+
+def collate():
+    #recreate the file to avoid crashes
+    Path(f'{snowglobes_dir}/fluxes/{fluence_name}.dat').touch()
+    try:
+        result = snowglobes.collate(snowglobes_dir,fluence_file, detector_input=detectors, 
+                                            skip_plots=True, remove_generated_files=False)
+    finally:
+        #return back to original directory
+        os.chdir(basedir)
+    return result
 
 #-------------------------------------------
 @pytest.mark.timing
@@ -75,6 +82,7 @@ def test_benchmark_simulate(benchmark):
 def test_benchmark_collate(benchmark):
     benchmark(collate)
 #-------------------------------------------
+@pytest.mark.snoglobes_steps
 def test_generate():
     #cleanup fluence file
     fluence_file.unlink(missing_ok=True)
@@ -83,15 +91,61 @@ def test_generate():
     assert result == str(fluence_file)
     assert fluence_file.exists()
 
+@pytest.mark.snoglobes_steps
+def test_simulate():
+    simulate()
+    #get files in output_dir
+    files = list(outdir.glob('{fluence_name}*.dat'))
+    #make sure we got many files
+    assert len(files)>1000
+    for f in files:
+        #and they're big enough
+        assert f.stat().size_t>100
+
+@pytest.mark.snoglobes_steps
 def test_collate():
     tables = collate()
     assert tables
+    files = list(outdir.glob('Collate_{fluence_name}*.dat'))
+    #make sure we got many files
+    assert len(files)>1000
+    for f in files:
+        #and they're big enough
+        assert f.stat().st_size>1024 #1KB
+#---------------------------------------------
+@pytest.mark.integration
+def test_run_snoglobes():
+    fluence_file.unlink(missing_ok=True)
+    generate()
+    simulate()
+    collate()
+    assert output_archive.exists()
+    assert output_archive.stat().st_size>102400 #100KB
 
-def test_output_files_match():
-    #check that each reference file has 
-    ref_files = list(Path(ref_dir).glob('*.dat'))
+
+#Prepare output directories for comparing archive files
+@pytest.fixture
+def temp_dirs(tmpdir):
+    dirs = [tmpdir/'ref', tmpdir/'new']
+    for d in dirs:
+        d.mkdir()
+    return dirs
+
+@pytest.mark.integration
+@pytest.mark.skipif(Path(reference_archive).exists() == False, reason='No reference file for check')
+def test_output_files_matches_reference(temp_dirs):
+    #check that each reference file has the same value
+    new_dir,ref_dir = temp_dirsz
+    #first extract all files
+    with tarfile.open(output_archive) as t:
+        t.extractall(new_dir)
+    with tarfile.open(reference_archive) as t:
+        t.extractall(ref_dir)
+    #now check that each file in the ref_dir is contained in the new_dir
+    ref_files = list(ref_dir.glob('*.dat'))
     assert len(ref_files)>0
     for ref_file in ref_files:
         if ref_file.stat().st_size>0:
-            _check_file_in_directory(ref_file, outdir)
+            _check_file_in_directory(ref_file, new_dir)
+
